afxdp/tx/cos_classify: relocate inline tests to cos_classify_tests.rs
Mirrors the worker/cos test relocation. tx/cos_classify.rs reaches
2,711 LOC with a 1,900-LOC `#[cfg(test)] mod tests { ... }` block at
the end (file becomes mostly tests).

Mechanical changes:
- Inline mod tests block (lines 811-2710) extracted to sibling
  cos_classify_tests.rs (1,903 lines incl 5-line header)
- 4-space dedent applied uniformly
- cos_classify.rs ends with:
    #[cfg(test)]
    #[path = "cos_classify_tests.rs"]
    mod tests;
- pre-rustfmt'd

cos_classify.rs drops from 2,711 → 814 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
